Use ShimLoader to access PlanShims (#6527)
