+++ sql-plugin/src/main/320+/scala/com/nvidia/spark/rapids/shims/PlanShimsImpl.scala
@@ -16,8 +16,13 @@
 
 package com.nvidia.spark.rapids.shims
 
-import org.apache.spark.sql.execution.SparkPlan
+import com.nvidia.spark.rapids.PlanShims
 
-object PlanShims {
-  def extractExecutedPlan(plan: SparkPlan): SparkPlan = plan
+import org.apache.spark.sql.execution.{CommandResultExec, SparkPlan}
+
+class PlanShimsImpl extends PlanShims {
+  def extractExecutedPlan(plan: SparkPlan): SparkPlan = plan match {
+    case p: CommandResultExec => p.commandPhysicalPlan
+    case _ => plan
+  }
 }
